fix: refactored shop files | fixed shop white screen render issue
